+++ app/static/lang/lang.es.js
@@ -51,6 +51,7 @@ export const lang = {
   surroundWithTagsText: { text: 'Envolver con tags' },
   surroundWithLastTagText: { text: 'Envolver con ' },
   jumpToLineText: { text: 'Ir a línea' },
+  toMatchingTagText: { text: 'Ir al tag correspondiente' },
   manualValidateText: { text: 'Validar' },
 
   // VIEW MENU ITEM / ELEMENTOS DEL MENÚ VER
@@ -310,15 +311,21 @@ export const lang = {
   selectTagNameForEnclosureCancelButton: { value: 'Cancelar' },
 
   // restore Solid session overlay
-  solidOverlayCancel: { html: 'Restaurando sesión de Solid - presione <span>Esc</span> o haga clic aquí para cancelar.' },
+  solidOverlayCancel: {
+    html: 'Restaurando sesión de Solid - presione <span>Esc</span> o haga clic aquí para cancelar.',
+  },
   solidWelcomeMsg: { text: 'Bienvenido, ' },
   solidLogout: { text: 'Cerrar sesión' },
+  solidLoggedOutWarning: {
+    html: `Has cerrado sesión en la integración Solid de mei-friend, ¡pero tu navegador todavía está iniciado en Solid!
+    <a id="solidIdPLogoutLink" target="_blank">Haz clic aquí para cerrar sesión en Solid</a>.`,
+  },
 
   // annotation panel / panel de anotaciones
   annotationCloseButtonText: { text: 'Cerrar panel de anotaciones' },
   hideAnnotationPanelButton: { description: 'Cerrar panel de anotaciones' },
   closeAnnotationPanelButton: { description: 'Cerrar panel de anotaciones' },
-  annotationToolsButton: { text: 'Herramientas', description: 'Herramientas de anotación' },
+  annotationToolsButton: { text: 'Utilidades', description: 'Herramientas de anotación' },
   annotationListButton: { text: 'Lista', description: 'Listar anotaciones' },
   writeAnnotStandoffText: { text: 'Anotación Web' },
   annotationToolsHighlightTitle: { text: 'Resaltar' },