+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 0.10.2
 * Improve UI around Solid handleIncomingRedirect
 * Provide translations for Solid UI
+* Set cursor position to matching tag name in editor
+* Bug that did not show elements within a choice in speed mode fixed
 
 ### 0.10.1
 * Fix contextual enabling / disabling of stand-off annotation interface